MealController override doGet() to show photo
