+++ common/src/main/java/com/common/QuickMealPhotoInsert.java
@@ -1,5 +1,8 @@
 package com.common;
-import java.io.*;
+import java.io.BufferedInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
@@ -9,9 +12,6 @@ import static com.common.DBConstants.USER;
 import static com.common.DBConstants.PASSWORD;
 public class QuickMealPhotoInsert {
     public static void main(String[] args) throws IOException {
-
-
-
         String INSERT_PHOTO = "UPDATE MEAL_PRODUCT SET MEAL_PHOTO = ? WHERE MEAL_NO = ? ;";
         for (int i = 1; i<=10 ; i++) {
             File file = new File("data/mealimages/"+i+".jpg");
@@ -25,8 +25,5 @@ public class QuickMealPhotoInsert {
                 e.printStackTrace();
             }
         }
-
-
-
     }
 }
