+++ common/src/main/java/com/meal/model/MealDAO.java
@@ -1,5 +1,6 @@
 package com.meal.model;
 
+import java.io.BufferedInputStream;
 import java.util.List;
 
 public interface MealDAO {
@@ -10,6 +11,7 @@ public interface MealDAO {
     MealVO findByMealName(String mealName);
     MealVO findWithoutAllergen(String mealAllergen);
 
+    BufferedInputStream showMealphoto(Integer mealNo);
     MealVO findByMealNo(Integer mealNo);
     List<MealVO> getAll();
 }
