+++ data/mealProductData.sql
@@ -10,4 +10,4 @@ VALUES
 ('增肌雞胸餐','1.舒肥雞胸 2.黑胡椒 3.鹽', 300,'無',150,'1.將產品室溫解凍 2.煎至兩面金黃調味',1),
 ('風味雞胸咖哩','1.咖哩調理包(含雞肉、胡蘿蔔、各式香料) 2.即食米飯',450,'無',150,'1.調理包隔水加熱 2.米飯微波2分鐘 3.混合即可食用',1),
 ('高纖風味凱薩沙拉','1.蘿蔓 2.番茄 3.起司 4.硬麵包 5.凱薩醬', 200,'起司',200,'1.將各式材料混勻即可食用',1),
-('香煎起司鮭魚','1.鮭魚片 2.起司 3.柳松菇、玉米筍(熟)',450,'起司',250,'1.將鮭魚煎至金黃 2.鋪上起司片以200度烤5分鐘 3.蔬菜裝飾',1)
+('香煎起司鮭魚','1.鮭魚片 2.起司 3.柳松菇、玉米筍(熟) 4.鹽',450,'起司',250,'1.將鮭魚煎至金黃 2.鋪上起司片以200度烤5分鐘 3.蔬菜裝飾',1)
